fix build warnings

+++ security/tlk_driver/ote_comms.c
@@ -598,7 +598,7 @@ static void do_smc_compat(struct te_request_compat *request,
 
 struct tlk_smc_work_args {
 	uint32_t arg0;
-	uint32_t arg1;
+	uintptr_t arg1;
 	uint32_t arg2;
 };
 
@@ -642,7 +642,7 @@ int te_set_vpr_params(void *vpr_base, size_t vpr_size)
 		int cpu = cpu_logical_map(smp_processor_id());
 
 		work_args.arg0 = TE_SMC_PROGRAM_VPR;
-		work_args.arg1 = (uint32_t)vpr_base;
+		work_args.arg1 = (uintptr_t)vpr_base;
 		work_args.arg2 = vpr_size;
 
 		/* workers don't change CPU. depending on the CPU, execute
